Renames <Mdx> to <Markdown>
The Mdx/ directory used to be mdx/ ( note the casing ) and changing to Mdx/ caused some issues with Vercel etc not picking it up.
